requirements.txt
chardet==3.0.4
pdfminer.six==20181108
mail-parser==3.9.3
beautifulsoup4==4.8.0
